pyproject.toml
[project]
name = "librhub"
version = "0.1.0"
description = "Blog and content sharing platform for developers"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.8",
    "pyjwt[crypto]>=2.10.1",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "sqlalchemy[asyncio]>=2.0.47",
    "uvicorn>=0.40.0",
    "argon2-cffi>=25.1.0",
    "slowapi>=0.1.9",
]
